BUG/MAJOR: api: Fix password treated unset for non-encrypted CA keys
Due to typecasting, the test "if (!req.session.caPassword)" returns
true when a CA password is set to an empty string, which is the case
when a CA key is not encrypted. As a consequence, the is-lock periodic
test coming from the client - as well as write operations on the CA -
will fail, causing the web ui to lock right after unlock.