BUG/MEDIUM: api: Fix broken session password expiration
Changing req.session.caPassword out of scope of the request (setTimeout)
does not work because the reference to req.session changes on each
request. Internally with the express' MemoryStore, session data is
serialized in JSON and gets recomputed (probably) on each request.
When the timer unsets req.session.caPassword, it does it on a reference
that was valid in the scope of the /set-password request. Thus, in fact
req.session.caPassword will never expire.

It is debatable how to work that around considering express-session
MemoryStore implementation, which commits session writes serialized in
JSON and parses it to load the data. The most reliable way would be to
store the expiration timestamp in the session and check it before using
the password (or clear it). For now we'll choose not to change the logic
and minimize the patch, we just change the timer to fetch, modify and
recommit the session data with a cleared password. It is most likely
racy if a request commits session data at same time, provided its session
data was loaded before the timer password clear commit. This should be
reviewed in future.