BUG/MEDIUM: zpki: Fix CRL never generated nor updated
When revoking a certificate, openssl ca was called with both -revoke
and -gencrl but the CRL file was never updated. There are two issues
here, test on an openssl 1.1.1

1) openssl ca will always generate the CRL (-gencrl) before revoking
certificate (-revoke), so the CRL would not be up-to-date.

2) openssl ca -gencrl does not automatically output to the path provided
in the crl parameter of the configuration file. I could not find a way
to make it working, the CRL is always printed on stdout, unless a -out
argument is given.