BUG/MEDIUM: api: Prevent zpki from waiting indefinitely on stdin
If a CA folder <olddir> is removed / renamed, the server can accumulate
zpki processes waiting on input for a user confirmation to answer the
question "Create CA storage directory '<olddir>' (y/n):".

The first thing is we can close stdin or pass /dev/null to it when
calling zpki. However zpki's yesno() function does not check read return
code and will go in an infinite loop. To fix that we can make it abort
if read fails. Also, since we can check if stdin is a terminal (there is
already a $HAS_TTY variable for that), we'd better just abort is YES=1
is not present in env.