MINOR: api: Review and optimize /profile and /switch-profile
Because the /profile handler can actually do a "switch profile" operation
there was duplicated code in both handlers; this is now refactored. In
order to minimize calls to the ca-folders command, we cache it for 5m by
default, which can be customized via CA_FOLDERS_CACHE_MS environment var.

If a CA folder disappear during cache time, zpki will just return an
error. If a folder is added during cache time, it won't be listed until
the cache expire.