MEDIUM: api: Do not regenerate the CSR on certificate renew
Certificate parameters (san, type) are currently not modifiable
from the web ui for a renew. So we do not need to regenerate
the CSR and thus do not need --force-csr. That way the private
key is not needed. Using --force-csr is problematic when the
private key is encrypted because the web ui does offer a way
enter the passphrase for it in the renew modal.